+++ Cargo.toml
@@ -29,6 +29,7 @@ solders-rpc-errors = { workspace = true }
 solders-rpc-requests = { workspace = true }
 solders-rpc-responses = { workspace = true }
 solders-pubkey = { workspace = true }
+solders-instruction = { workspace = true }
 solders-transaction-status = { workspace = true }
 solders-bankrun = { workspace = true, optional = true }
 
@@ -58,6 +59,7 @@ solders-rpc-requests = { path = "./crates/rpc-requests", version = "0.14.4" }
 solders-rpc-responses = { path = "./crates/rpc-responses", version = "0.14.4" }
 solders-primitives = { path = "./crates/primitives", version = "0.14.4" }
 solders-pubkey = { path = "./crates/pubkey", version = "0.14.4" }
+solders-instruction = { path = "./crates/instruction", version = "0.14.4" }
 solders-account = { path = "./crates/account", version = "0.14.4" }
 serde_with = { version = "^1.14.0", features = ["base64"] }
 solana-sdk = "^1.15.2"
