+++ crates/instruction/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solders-primitives"
+name = "solders-instruction"
 version = "0.14.4"
 edition = "2021"
 include = ["/src"]
@@ -12,10 +12,6 @@ pyo3 = { workspace = true, features = ["macros"] }
 solders-traits = { workspace = true }
 solders-macros = { workspace = true }
 solders-pubkey = { workspace = true }
-solders-instruction = { workspace = true }
 solana-sdk = { workspace = true }
 serde = { workspace = true }
-base64 = { workspace = true }
 derive_more = { workspace = true }
-serde_with = { workspace = true }
-bincode = { workspace = true }
