+++ crates/primitives/src/transaction.rs
@@ -12,7 +12,7 @@ use solana_sdk::{
         VersionedTransaction as VersionedTransactionOriginal,
     },
 };
-use solders_pubkey::Pubkey;
+use solders_pubkey::{Pubkey, convert_optional_pubkey};
 use solders_macros::{common_methods, richcmp_eq_only, EnumIntoPy};
 use solders_traits::{
     handle_py_err, impl_display, py_from_bytes_general_via_bincode, pybytes_general_via_bincode,
@@ -20,15 +20,14 @@ use solders_traits::{
 };
 
 use crate::{
-    convert_instructions, convert_optional_pubkey,
     hash::Hash as SolderHash,
-    instruction::{CompiledInstruction, Instruction},
     message::{Message, VersionedMessage},
     signature::Signature,
     signature::{originals_into_solders, solders_into_originals},
     signer::Signer,
     signer::SignerVec,
 };
+use solders_instruction::{convert_instructions, CompiledInstruction, Instruction};
 
 /// An atomic transaction
 ///
