+++ src/lib.rs
@@ -5,7 +5,7 @@ use address_lookup_table_account::create_address_lookup_table_account_mod;
 use pyo3::prelude::*;
 use rpc::create_rpc_mod;
 use solders_account::create_account_mod;
-use solders_primitives::instruction::{AccountMeta, CompiledInstruction, Instruction};
+use solders_instruction::{AccountMeta, CompiledInstruction, Instruction};
 use solders_traits::{BincodeError, CborError, ParseHashError, SerdeJSONError, SignerError};
 use solders_transaction_status::create_transaction_status_mod;
 use std::collections::HashMap;
